Closes #1: Start votes at 0
Previously comments and topics started with a value of 1, to which a user could add an additional vote - whether their own comment or topic or otherwise. Now we start votes at 0, so that it feels less strange that you upvote yourself to get to 1 vote. In the future, creating a topic or comment may automatically submit a vote for your new topic or comment, and this commit lays the foundation for that change.